+++ src/main/java/seng201/team53/items/Purchasable.java
@@ -5,7 +5,7 @@ public interface Purchasable {
 
     String getDescription();
 
-    Double getCostPrice();
+    int getCostPrice();
 
     Double getSellPrice();
