+++ src/main/java/seng201/team53/items/towers/WindMillTower.java
@@ -15,8 +15,8 @@ public class WindMillTower extends Tower {
     }
 
     @Override
-    public Double getCostPrice() {
-        return 100.0;
+    public int getCostPrice() {
+        return 100;
     }
 
     @Override
